Unit tests were added for MarkDao (not all are green)

+++ faculty/src/main/java/com/ra/course/janus/faculty/dao/MarkDao.java
@@ -13,7 +13,7 @@ public interface MarkDao {
 
     boolean delete(Mark mark) throws SQLException;
 
-    Mark findByTid(Integer tid) throws SQLException;
+    Mark findByTid(long tid) throws SQLException;
 
     List<Mark> findAll() throws SQLException;
 }
